+++ src/types/soap/index.d.ts
@@ -17,3 +17,4 @@ export * from './sync';
 export * from './zimbra-request';
 export * from './save-draft';
 export * from './redirect-message-action';
+export * from './create-smart-links';
